+++ src/main/java/TelegramBot/task/TaskData.java
@@ -9,7 +9,7 @@ import java.sql.Timestamp;
 @Setter
 @Getter
 public class TaskData {
-    private long dbID;
+    private int dbID;
     private long chatId;
     private String description;
     private Timestamp deadline;
@@ -25,7 +25,7 @@ public class TaskData {
     public TaskData(Long chatId) {
         this.chatId = chatId;
     }
-    public TaskData(Long dbID, Long chatId, String description, Timestamp deadline, String priority, String status, Timestamp creationDate) {
+    public TaskData(int dbID, Long chatId, String description, Timestamp deadline, String priority, String status, Timestamp creationDate) {
         this.dbID = dbID;
         this.chatId = chatId;
         this.description = description;
